Add notifications

+++ utilities/rename-random.py
@@ -16,6 +16,8 @@ import sys
 
 
 def set_working_root():
+    print(f"Checking system")
+
     # Handle windows and cygwin paths 
     system_os = platform.system()
     root = ""
@@ -36,6 +38,8 @@ def change_directories(target):
 
 
 def rename_files():
+    print(f"Renaming Files")
+
     new_names = set()
     original_files = []
 
@@ -62,41 +66,51 @@ def rename_files():
             print("{}: {}".format(of, e))
 
 
-def remove_old_set():
+def remove_old_set(root, target_dir):
     # Remove the old working dir and files
-    fotos_dir = os.path.join(WORKING_ROOT, WORKING)
-    try: 
-        os.chdir(fotos_dir)
-        for f in os.listdir():
-            try: 
-               os.remove(f)
-            except FileNotFoundError as e:
-               print(f"Can't remove {f}, error: {e}")
-        os.chdir("..")
-        os.rmdir(WORKING)
-    except Exception as outer_e:
-       print(f"Can't find directory {fotos_dir}, error: {outer_e}")
+    fotos_dir = os.path.join(root, target_dir)  
+    print(f"Removing files")
+    
+    if os.path.isdir(fotos_dir):
+        try: 
+            print(f"     Removing files from: {fotos_dir}")
+            os.chdir(fotos_dir)
+            for f in os.listdir():
+                try: 
+                   os.remove(f)
+                except FileNotFoundError as e:
+                   print(f"Can't remove {f}, error: {e}")
+            os.chdir("..")
+            print(f"     Removing directory: {target_dir}")
+            os.rmdir(target_dir)
+        except Exception as outer_e:
+            print(f"Can't cd to directory {fotos_dir}, error: {outer_e}")
+    else:
+        print(f"Can't find directory {fotos_dir}")
 
 
-def create_new_set():
+def create_new_set(root, target_dir):
     # Create a new working directory
+    fotos_dir = os.path.join(root, target_dir)
+    print(f"Creating directory: {fotos_dir}")
+
     try: 
-        fotos_dir = os.path.join(WORKING_ROOT, WORKING)
         os.mkdir(fotos_dir)
     except OSError as e:
-        print(f"Can't create directory: {WORKING} error: {e}")   
-        sys.exit()
-
+        print(f"Can't create directory: {target_dir} error: {e}")   
 
     # Move about 10% of files to the new directory    
+    print(f"Populating directory: {fotos_dir}")
+
     for entry in os.listdir():
         if os.path.isfile(entry):    
             move = random.randint(0,10)
             if move == 0: 
                 try:
-                    shutil.move(entry, WORKING)
+                    shutil.move(entry, target_dir)
                 except OSError as move_e:
                     print(f"Can't move: {entry}, error: {move_e}")
+
                 
 
 if __name__ == "__main__":
@@ -106,4 +120,6 @@ if __name__ == "__main__":
     WORKING_ROOT = set_working_root()
     change_directories(WORKING_ROOT)
     rename_files()
+    remove_old_set(WORKING_ROOT, WORKING)
+    create_new_set(WORKING_ROOT, WORKING)
